resource_builders.py file moved out of the utils folder to follow the convention in the codebase. resource_builders.py file renamed to utils.py file. get_model_by_name moved into the utils.py file.
