+++ cms/search/serializers.py
@@ -2,7 +2,7 @@ from typing import TYPE_CHECKING
 
 from rest_framework import serializers
 
-from cms.search.utils.resource_builders import build_resource_dict
+from cms.search.utils import build_resource_dict
 
 if TYPE_CHECKING:
     from wagtail.models import Page
