+++ cms/search/views.py
@@ -1,6 +1,5 @@
 from typing import TYPE_CHECKING
 
-from django.apps import apps
 from django.conf import settings
 from rest_framework import status
 from rest_framework.response import Response
@@ -11,6 +10,7 @@ from cms.settings.base import SEARCH_INDEX_EXCLUDED_PAGE_TYPES
 
 from .pagination import CustomPageNumberPagination
 from .serializers import ResourceSerializer
+from .utils import get_model_by_name
 
 if TYPE_CHECKING:
     from django.http import HttpRequest
@@ -40,18 +40,11 @@ class ResourceListView(APIView):
 
         return paginator.get_paginated_response(data)
 
-    def get_model_by_name(self, model_name: str) -> type:
-        # Loop through all installed models in all apps
-        for model in apps.get_models():
-            if model.__name__ == model_name:
-                return model
-        raise LookupError(f"No model named '{model_name}' was found.")
-
     def get_queryset(self) -> list[Page]:
         """Returns a queryset of 'published' pages that are indexable,
         excluding pages we do not want to index.
         """
-        excluded_classes = [self.get_model_by_name(name) for name in SEARCH_INDEX_EXCLUDED_PAGE_TYPES]
+        excluded_classes = [get_model_by_name(name) for name in SEARCH_INDEX_EXCLUDED_PAGE_TYPES]
         qs: list[Page] = Page.objects.live().public().not_exact_type(*excluded_classes).specific().defer_streamfields()
 
         return qs
